Storm: Output filename is now optional, coords-File is always saved
If no output filename is given, the result is saved in the same folder the input data is in. The image is saved as PNG with the same name with file extension .sif substituted by .png. The same holds for coordinates file: If no name is given, it is saved with the filename substitued by .txt